+++ app/src/main/java/nextstep/github/data/api/model/RepositoryEntity.kt
@@ -1,0 +1,12 @@
+package nextstep.github.data.api.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RepositoryEntity(
+    @SerialName("full_name")
+    val fullName: String?,
+    @SerialName("description")
+    val description: String?
+)
