--- app/src/main/java/nextstep/github/network/RepositoryEntity.kt
@@ -1,10 +1,0 @@
-package nextstep.github.network
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RepositoryEntity(
-    @SerialName("full_name") val fullName: String?,
-    @SerialName("description") val description: String?
-)
